Create tracking service package and move classes into it
