+++ app/src/main/java/com/crankworks/trackingservice/DummyRecorder.java
@@ -1,4 +1,4 @@
-package com.crankworks.crankanonymous;
+package com.crankworks.trackingservice;
 
 /**
  * Created by marcus on 12/7/14.
