+++ app/src/main/java/com/crankworks/trackingservice/RecorderStateIdle.java
@@ -1,11 +1,11 @@
-package com.crankworks.crankanonymous;
+package com.crankworks.trackingservice;
 
 /**
  * Created by marcus on 12/8/14.
  */
-public class RecorderStatePause extends RecorderStateBase implements IRecorder
+public class RecorderStateIdle extends RecorderStateBase implements IRecorder
 {
-    public RecorderStatePause(TrackingRecorder stateContext)
+    public RecorderStateIdle(TrackingRecorder stateContext)
     {
         super(stateContext);
     }
@@ -17,11 +17,9 @@ public class RecorderStatePause extends RecorderStateBase implements IRecorder
 
     public void finishRecording()
     {
-        getListener().recorderIdle();
     }
 
     public void cancelRecording()
     {
-        getListener().recorderIdle();
     }
 }
