+++ app/src/main/java/com/crankworks/trackingservice/TrackingRecorder.java
@@ -1,11 +1,8 @@
-package com.crankworks.crankanonymous;
+package com.crankworks.trackingservice;
 
 import android.content.Context;
 import android.location.Criteria;
-import android.location.Location;
-import android.location.LocationListener;
 import android.location.LocationManager;
-import android.os.Bundle;
 import android.util.Log;
 
 /**
