+++ app/src/main/java/com/crankworks/trackingservice/TrackingService.java
@@ -1,4 +1,4 @@
-package com.crankworks.crankanonymous;
+package com.crankworks.trackingservice;
 
 import android.app.Service;
 import android.content.Intent;
